[#106731208] Refactored the methods name to reflect REST API naming convention
